requirements.txt
telethon
python-dotenv
tqdm
psycopg2-binary
pandas
sqlalchemy